+++ backend/src/main/java/com/pingpong/backend/api/repository/TimetableRepository.java
@@ -4,5 +4,5 @@ package com.pingpong.backend.api.repository;
 import com.pingpong.backend.api.domain.TimetableEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TimetableRepository extends JpaRepository<TimetableEntity, Byte> {
+public interface TimetableRepository extends JpaRepository<TimetableEntity, Integer> {
 }
